requirements.txt
# beautifulsoup4==4.8.1
# certifi==2019.11.28
# cffi==1.13.2
# chardet==3.0.4
# Click==7.0
# cryptography==2.8
# future==0.18.2
# google-images-download==2.8.0
# gTTS==2.0.4
# gTTS-token==1.1.3
# idna==2.8
# pycparser==2.19
# pymongo==3.9.0
# python-git-info==0.6.1
python-telegram-bot==12.2.0
# requests==2.22.0
# selenium==3.141.0
# six==1.13.0
# soupsieve==1.9.5
# tinydb==3.15.1
# tornado==6.0.3